pyproject.toml
[tool.poetry]
name = "nonebot-plugin-lingyi-chat"
version = "0.1.4"
description = ""
authors = ["Yeying-Xingchen <yeyingxingchen@yeah.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
nonebot2 = "^2.2.0"
aiohttp = "^3.10.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
